+++ src/models/vs-code-position-shim.ts
@@ -3,11 +3,5 @@
 import { IVisualCodePositionShim } from './interfaces/vs-code-position-shim';
 
 export class VisualCodePositionShim implements IVisualCodePositionShim {
-    readonly line: number;
-    readonly character: number;    
-
-    constructor(line: number, character: number) {
-        this.line = line;
-        this.character = character;
-    }
+    constructor(public readonly line: number, public readonly character: number) {}
 }